+++ src/Function/AbstractFunction.php
@@ -23,11 +23,6 @@ abstract class AbstractFunction implements FunctionInterface
         $this->setTable($table);
     }
 
-    public function execute(?array $params = null): bool|array|object
-    {
-        return $this->getConnection()->execute($this->getQueryString(), $params);
-    }
-
     public function arguments(mixed ...$arguments): static
     {
         $this->arguments = [];